requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
seaborn
pillow
flask
flask-cors